+++ server/routes/db_config.js
@@ -13,13 +13,22 @@ if(env === 'development') {
       database: process.env.DB_NAME,
       password: process.env.DB_PASSWORD,
       port: process.env.DB_PORT,
-    }
+    };
+}
+else if(env === 'test') {
+  connection = {
+    user: process.env.DB_USER,
+    host: process.env.DB_HOST,
+    database: process.env.DB_NAME,
+    password: process.env.DB_PASSWORD,
+    port: process.env.DB_PORT,
+  };
 }
 else {
   connection = {
     connectionString: process.env.DATABASE_URL,
     ssl: true
-  }
+  };
 }
 
 const pool = new Pool(connection);
